refactor(cloud): the staging dispatch moves to where the agent modules are
Phase 4b, step one. `stageAllAgentStatic` lived in `sandbox-core` and reached
its per-agent stagers through a `DEDICATED_STATIC_STAGERS` table declared right
beside them — which is exactly what pinned those stagers in `sandbox-core`. An
agent package depends on `sandbox-core`, so `sandbox-core` cannot import one
back.

The dispatch now lives in `sandbox-cloud` and reads `AgentCloudModule.stageStatic`
— a seam that already existed, on a contract that already had the method. What
stays in `sandbox-core` is the part with no agent in it: the generic
`stageAgentStaticForUpload`, which stages any agent from its `staticPaths`
alone, plus the rsync/tar primitives.

`stageAllAgentStatic` is a cloud-staging concern anyway: every consumer is a
cloud provider's `prepare` or the provider SDK. All five providers and the SDK
now import it from `@agentbox/sandbox-cloud`; `pack:test` still passes with the
same 31 exports.

Its test moved with it, because it has to — the dispatch reads a registry that
only exists in this package.
