+++ packages/sandbox-core/src/sync/index.ts
@@ -180,7 +180,8 @@ export {
   stageOpencodeStaticForUpload,
   stageOpencodeCredentialsForUpload,
   stageOpencodeStateForUpload,
-  stageAllAgentStatic,
+  stageAgentStaticForUpload,
+  AGENTS_STATIC_BOX_DIR,
   type AgentStaticStage,
   type StageClaudeOptions,
   type StageCodexOptions,
