+++ packages/sandbox-core/test/host-stage.test.ts
@@ -5,10 +5,8 @@ import { join } from 'node:path';
 import { execa } from 'execa';
 import {
   stageAgentStaticForUpload,
-  stageAllAgentStatic,
   stageOpencodeStaticForUpload,
 } from '../src/sync/host-stage.js';
-import { AGENT_SYNC_SPECS } from '../src/sync/registry.js';
 
 /** Every path inside a staged tarball, relative and without the leading `./`. */
 async function tarEntries(tarball: string): Promise<string[]> {
@@ -95,31 +93,6 @@ describe('stageAgentStaticForUpload', () => {
         expect(res.tarballPath).toBeNull();
         await res.cleanup();
       } finally {
-        await rm(home, { recursive: true, force: true });
-      }
-    },
-    SUBPROCESS_TIMEOUT_MS,
-  );
-});
-
-describe('stageAllAgentStatic', () => {
-  it(
-    'covers every registered agent, so a new one reaches the cloud snapshots',
-    async () => {
-      // Scoped to the demo agent against an empty fixture home: the real
-      // claude/codex stagers would rsync the developer's own `~/.claude`, which
-      // is neither cheap nor deterministic. What matters here is the DISPATCH —
-      // that the set comes from the registry, not from a hardcoded triple.
-      const home = await mkdtemp(join(tmpdir(), 'agentbox-stage-test-'));
-      const stages = await stageAllAgentStatic({ agents: ['example'], hostHome: home });
-      try {
-        expect(stages.map((s) => s.kind)).toEqual(['example', 'agents']);
-        const example = stages.find((s) => s.kind === 'example');
-        expect(example?.extractDir).toBe(
-          AGENT_SYNC_SPECS.find((s) => s.id === 'example')?.staticPaths[0]?.boxDir,
-        );
-      } finally {
-        for (const s of stages) await s.staged.cleanup();
         await rm(home, { recursive: true, force: true });
       }
     },
